- Added Macro for formatting any large long or int numbers - removed FormatByteSizeMacro as it was duplicate of FormatSizeMacro
